move hit testing/hover to base
