+++ bokehjs/src/lib/models/glyphs/image.ts
@@ -6,19 +6,10 @@ import {Arrayable} from "core/types"
 import * as p from "core/properties"
 import {concat} from "core/util/array"
 import {Context2d} from "core/util/canvas"
-import * as hittest from "core/hittest"
-import {Selection} from "../selections/selection"
-import {PointGeometry} from "core/geometry"
-import { ImageRGBAData } from './image_rgba'
 
-export interface ImageView extends ImageDataBase {}
+export interface ImageData extends ImageDataBase {}
 
-export interface ImageIndex {
-  index: number
-  dim1: number
-  dim2: number
-  flat_index: number
-}
+export interface ImageView extends ImageData {}
 
 export class ImageView extends ImageBaseView {
   model: Image
@@ -39,34 +30,6 @@ export class ImageView extends ImageBaseView {
       this._set_data()
       this.renderer.plot_view.request_render()
     }
-  }
-
-  _image_index(index : number, x: number, y : number) : ImageIndex {
-    const [l,r,t,b] = this._lrtb(index)
-    const width = this._width[index]
-    const height = this._height[index]
-    const dx = (r - l) / width
-    const dy = (t - b) / height
-    const dim1 = Math.floor((x - l) / dx)
-    const dim2 = Math.floor((y - b) / dy)
-    return {index, dim1, dim2, flat_index: dim2*width + dim1}
-  }
-
-  _hit_point(geometry: PointGeometry) : Selection {
-    const {sx, sy} = geometry
-    const x = this.renderer.xscale.invert(sx)
-    const y = this.renderer.yscale.invert(sy)
-    const bbox = hittest.validate_bbox_coords([x, x], [y, y])
-    const candidates = this.index.indices(bbox)
-    const result = hittest.create_empty_hit_test_result()
-
-    result.image_indices = []
-    for (const index of candidates) {
-      if ((sx != Infinity) && (sy != Infinity)) {
-        result.image_indices.push(this._image_index(index, x,y))
-      }
-    }
-    return result
   }
 
   protected _set_data(): void {
@@ -94,7 +57,7 @@ export class ImageView extends ImageBaseView {
     }
   }
 
-  protected _render(ctx: Context2d, indices: number[], {image_data, sx, sy, sw, sh}: ImageRGBAData): void {
+  protected _render(ctx: Context2d, indices: number[], {image_data, sx, sy, sw, sh}: ImageData): void {
     const old_smoothing = ctx.getImageSmoothingEnabled()
     ctx.setImageSmoothingEnabled(false)
